requirements.txt
fastapi
uvicorn
pydantic
openai
pinecone-client
python-dotenv